Add basic Texture wrapper and Vertex struct
This adds a basic wrapper for the `WebGlTexture` object and includes some sample usage/visual tests in `lib.rs`. New shaders for basic texture rendering are included

Vertex is a struct for bundling together vertex data (position, uv coords). This may later be split into multiple Vertex types since the 3D renderer will need normals as well if lighting is included in the simulation

An empty struct for the 2D renderer is in this commit as well