Add resource manager prototype
This adds some prototype code for a resource manager. It works, but I'm not sure that downcasting is necessarily a good/efficient way to do this. Also, the majority of lib.rs is currently commented out since making gl_object private to the module breaks some things that I'm going to fix as I continue testing/implementing