Add macro for generating custom managers
This adds a macro to implement custom managers (i.e., GlObjectManager). It takes a type name that will be used for function and variable names and a type that is used to actually use as the type for. Multiple types can be added to the custom manager in this way.

This is still experimental and very much a WIP; it isn't actually being used or tested anywhere yet