Add impl_buffer macro & successful compilation
This adds a impl_buffer macro that implements the Buffer trait on a given struct or struct name. It allows for easy creation of buffer types that are essentially copy-paste clones of each other with different names (i.e. array buffers and element array buffers) as well as creating a buffer implementor that requires additional fields.

This also changes the define_manager macro to allow for specific paths to be specified for managed structs

This commit doesn't include the actual use of the new manager and buffer system in runtime, however the program compiles successfully.

Related to #17