Integrate GlObjectManager
This integrates the new GlObjectManager system into the rest of the program, aka: all GlObjects are stored within the manager and references to the manager are passed around rather than individual GlObjects.

This new system owns all of the GlObjects that it contains and tracks the currently bound object for each type. References to GlObjects can be retrieved with manager.get_* or manager.get_mut_*, where * is
the struct name in snake_case (i.e. ShaderProgram is get_shader_program and get_mut_shader_program). manager.reload() also reloads all GlObjects within and restores the currently bound objects

Related to #17, #16, and #13