Add onload response buffer and arg structs
This makes the request return the response as a Arraybuffer of bytes and passes it to the given callback.

This adds CallbackArgs and OnloadCallbackArgs structs to allow `_` in anon func parameters when setting the callbacks for a request

This adds allow(dead_code) to all set_* functions in ResourceLoader

ResourceLoader now uses `into()` to convert closure JsValues into a Function rather than `dyn_into()`. It also now checks the return value of the JS function call for errors