Integrate ResourceLoader and ResourceManager
This integrates ResourceLoader and ResourceManager into the rest of the program. init() is where resources are loaded, and start() is called after all resources are available within the resource manager. The cube obj is being loaded through the resource loader as well as the shaders. Consequently, the embedded cube OBJ file and shader source code has been removed. Renderer now pulls the shader source from the resource manager and reads it into a string before creating a new shader program.