Fix canvas sizing bug
This adds some javascript callbacks for window resizing and onload to set the width and height attributes of the canvas element. On window resize, the WebGL context viewport is set to be a 16:9 resolution that is the closest to the new size of the canvas. This is to prevent any image distortion